wip on non-ssl raw socket client
Bug: crashpad:196, crashpad:227

+++ util/net/http_transport_test_server.cc
@@ -77,7 +77,7 @@ int HttpTransportTestServerMain(int argc, char* argv[]) {
              svr.stop();
            });
 
-  int port = svr.bind_to_any_port("127.0.0.1");
+  uint16_t port = svr.bind_to_any_port("127.0.0.1");
 
   CheckedWriteFile(
       StdioFileHandle(StdioStream::kStandardOutput), &port, sizeof(port));
